Allow users to read other user objects when both users share an org
This changes the creation policy for newly created organizations and
provides a migration for existing organizations to ensure that the
users group is added to the ORGNAME_global_admins group.

The ORGNAME_global_admins group has one purposes inside the Chef
Server: to provide READ access to users that are associated with
ORGNAME. Previously only admins were granted this READ access.
However, users of chef-vault would like ordinary users to be able to
encrypt items for other users, thus promping this change in behavior.

The following API requests that would have previously failed, will now
succeed if two users are members of the same organization:

- GET /users/USERNAME
- GET /users/USERNAME/organizations
- GET /users/USERNAME/keys
- GET /users/USERNAME/keys/KEYNAME